main/agentforge-engine/pyproject.toml
[project]
name = "pyagentforge"
version = "3.0.0"
description = "通用型 AI Agent 服务底座 - 模型即代理，代码即配置"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "PyAgentForge Team", email = "team@pyagentforge.dev" }
]
keywords = ["ai", "agent", "llm", "claude", "openai", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.10.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.0",
    "types-passlib>=1.7.0",
]
redis = [
    "redis>=5.0.0",
]
postgres = [
    "asyncpg>=0.29.0",
]
all = [
    "pyagentforge[dev,redis,postgres]",
]

# ---------------------------------------------------------------------------
# Protocol adapter 插件入口点
# ---------------------------------------------------------------------------
# 任何第三方包只需在自己的 pyproject.toml 中声明：
#
#   [project.entry-points."pyagentforge.protocol_adapters"]
#   my-vendor = "my_pkg.adapter:MyVendorProtocol"
#
# 即可被 kernel 在 import ``pyagentforge.protocols`` 时自动发现并注册，
# 无需修改 kernel 源码或手动调用 ``register_protocol_adapter``。
# 下面登记的是随 kernel 一起分发的 "内置 vendor"，它们位于
# ``pyagentforge.plugins.protocol.*`` 子包下，与第三方插件走完全相同的
# 装配通道。当包未通过 pip 安装（纯源码开发）时，协议模块会走 bundled
# fallback 继续保证可用。
[project.entry-points."pyagentforge.protocol_adapters"]
openai-completions = "pyagentforge.plugins.protocol.openai.chat:OpenAIChatProtocol"
openai-responses = "pyagentforge.plugins.protocol.openai.responses:OpenAIResponsesProtocol"
anthropic-messages = "pyagentforge.plugins.protocol.anthropic.messages:AnthropicMessagesProtocol"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pyagentforge"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip without explicit strict
    "ARG001", # unused function argument
]

[tool.ruff.lint.isort]
known-first-party = ["pyagentforge"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG002", "ARG005", "F401", "F403", "E741", "E402", "B017", "B018", "SIM105"]
"examples/**/*.py" = ["ARG002", "ARG005", "F401", "F403", "E741", "E402", "B017", "B018", "SIM105"]
"pyagentforge/**/tests/**/*.py" = ["ARG002", "ARG005", "F401", "F403", "E741", "E402", "B017", "B018", "SIM105"]
"pyagentforge/plugins/**/*.py" = ["ARG002", "ARG005", "SIM102", "B007"]
"pyagentforge/plugins/**/examples/__init__.py" = ["F403"]
"pyagentforge/plugins/**/tests/__init__.py" = ["F403"]
"pyagentforge/plugin/base.py" = ["B027"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
exclude = [
    '^pyagentforge/.*/tests/',
]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "apscheduler.*",
    "slowapi.*",
    "websockets.*",
    "watchdog.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "pyagentforge.*.tests.*",
    "pyagentforge.plugins.*.tests.*",
]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "pyagentforge.plugins.*",
    "pyagentforge.plugin.*",
    "pyagentforge.codesearch.*",
    "pyagentforge.capabilities.*",
    "pyagentforge.middleware.*",
    "pyagentforge.tools.*",
    "pyagentforge.prompts.*",
    "pyagentforge.skills.*",
    "pyagentforge.workflow.*",
    "pyagentforge.automation.*",
    "pyagentforge.lsp.*",
    "pyagentforge.mcp.*",
    "pyagentforge.protocols",
]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short -p no:cacheprovider"

[tool.coverage.run]
source = ["pyagentforge"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
